requirements.txt
discord.py>=2.3.0
openai>=1.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
aiosqlite>=0.17.0
asyncio
requests>=2.28.0
python-dateutil>=2.8.0
Pillow>=9.0.0
aiofiles>=0.8.0
psutil>=5.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
yt-dlp>=2023.12.30
PyNaCl>=1.5.0
pytz>=2023.3.0
openai>=1.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
aiosqlite>=0.17.0
asyncio
requests>=2.28.0
python-dateutil>=2.8.0
Pillow>=9.0.0
aiofiles>=0.8.0
psutil>=5.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
yt-dlp>=2023.12.30
PyNaCl>=1.5.0
